[updated] OR
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk/MMOCoreORB@1641 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/packets/creature/CreatureObjectMessage3.h
@@ -60,7 +60,7 @@ public:
 		insertLong(creo->getCreatureLinkID()); // creature link id/ mount
 
 		insertFloat(creo->getHeight());
-		insertInt(creo->getShockWounds()); // BF
+		insertInt((int)creo->getShockWounds()); // BF
 
 		insertLong(creo->getStateBitmask());
